Unify form and form builder, rename form handler
