+++ crates/brace-web-auth/src/lib/form/login.rs
@@ -1,14 +1,14 @@
-use brace_web_form::{field, FormBuilder, FormHandler};
+use brace_web_form::{field, Form, FormBuilder};
 use failure::Error;
 use uuid::Uuid;
 
 pub struct LoginForm;
 
-impl FormHandler for LoginForm {
+impl FormBuilder for LoginForm {
     type Context = ();
-    type Future = Result<FormBuilder, Error>;
+    type Future = Result<Form, Error>;
 
-    fn build(&self, mut form: FormBuilder, _: Self::Context) -> Self::Future {
+    fn build(&self, mut form: Form, _: Self::Context) -> Self::Future {
         form.insert(field::hidden("id").value(Uuid::new_v4().to_string()));
 
         form.insert(
