+++ crates/brace-web-auth/src/lib/form/user.rs
@@ -1,14 +1,14 @@
-use brace_web_form::{field, FormBuilder, FormHandler};
+use brace_web_form::{field, Form, FormBuilder};
 use chrono::{DateTime, NaiveDateTime, Utc};
 use failure::Error;
 
 pub struct UserForm;
 
-impl FormHandler for UserForm {
+impl FormBuilder for UserForm {
     type Context = ();
-    type Future = Result<FormBuilder, Error>;
+    type Future = Result<Form, Error>;
 
-    fn build(&self, mut form: FormBuilder, _: Self::Context) -> Self::Future {
+    fn build(&self, mut form: Form, _: Self::Context) -> Self::Future {
         form.insert(field::hidden("id").value(form.state().get::<String>("id")?));
 
         form.insert(
